feat: Add live connection support and refactor related components
